add header

+++ index.js
@@ -1,20 +1,23 @@
-const Discord = require("discord.js");
-const bot = new Discord.Client({ intents: [Discord.Intents.FLAGS.GUILDS, Discord.Intents.FLAGS.GUILD_MEMBERS, Discord.Intents.FLAGS.DIRECT_MESSAGES, Discord.Intents.FLAGS.GUILD_MESSAGES] });
+const Discord = require('discord.js');
+const bot = new Discord.Client({
+  intents: [Discord.Intents.FLAGS.GUILDS, Discord.Intents.FLAGS.GUILD_MEMBERS, Discord.Intents.FLAGS.DIRECT_MESSAGES, Discord.Intents.FLAGS.GUILD_MESSAGES]
+});
+bot.commands = new Discord.Collection();
 
 const config = require('./config.json');
 const fivem = require('./server/info.js');
-const { Pagination } = require("discordjs-button-embed-pagination");
+const fs = require('fs');
+const commandFiles = fs.readdirSync('./commands/').filter(file => file.endsWith('.js'));
 
 const PREFIX = config.PREFIX;
-const PERMISSION = 'MANAGE_MESSAGES';
 const COLORBOX = config.COLORBOX;
 const NAMELIST = config.NAMELIST;
 const NAMELISTENABLE = config.NAMELISTENABLE;
-const AUTODELETE = config.AUTODELETE;
 const SERVER_NAME = config.SERVER_NAME;
 const BOT_TOKEN = config.BOT_TOKEN;
 const UPDATE_TIME = config.UPDATE_TIME;
 const SERVER_LOGO = config.SERVER_LOGO;
+const AUTODELETE = config.AUTODELETE;
 
 const inFo = new fivem.ApiFiveM(config.URL_SERVER);
 
@@ -23,7 +26,9 @@ var STATUS;
 console.logCopy = console.log.bind(console);
 console.log = function (data) {
   var timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-  var currentDate = '|' + new Date().toLocaleString({ timeZone: timezone }).slice(10, -3) + '|';
+  var currentDate = '|' + new Date().toLocaleString({
+    timeZone: timezone
+  }).slice(10, -3) + '|';
   this.logCopy(currentDate, data);
 };
 
@@ -38,20 +43,28 @@ const activity = async () => {
       });
 
       if (playersonline === 0) {
-        bot.user.setActivity(`⚠ Wait for Connect`, { 'type': 'WATCHING' });
+        bot.user.setActivity(`⚠ Wait for Connect`, {
+          'type': 'WATCHING'
+        });
         console.log(`Wait for Connect update at activity`);
       } else if (playersonline >= 1) {
         if (NAMELISTENABLE) {
-          bot.user.setActivity(`💨 ${playersonline}/${maxplayers} 👮‍ ${namef.length}`, { 'type': 'WATCHING' });
+          bot.user.setActivity(`💨 ${playersonline}/${maxplayers} 👮‍ ${namef.length}`, {
+            'type': 'WATCHING'
+          });
           console.log(`Update ${playersonline} at activity`);
         } else {
-          bot.user.setActivity(`💨 ${playersonline}/${maxplayers}`, { 'type': 'WATCHING' });
+          bot.user.setActivity(`💨 ${playersonline}/${maxplayers}`, {
+            'type': 'WATCHING'
+          });
           console.log(`Update ${playersonline} at activity`);
         }
       }
 
     } else {
-      bot.user.setActivity(`🔴 Offline`, { 'type': 'WATCHING' });
+      bot.user.setActivity(`🔴 Offline`, {
+        'type': 'WATCHING'
+      });
       console.log(`Offline at activity`);
     }
 
@@ -59,28 +72,6 @@ const activity = async () => {
     console.log(`Catch ERROR` + err);
   });
 };
-
-function splitChunks(sourceArray, chunkSize) {
-  let result = [];
-  for (var i = 0; i < sourceArray.length; i += chunkSize) {
-    result[i / chunkSize] = sourceArray.slice(i, i + chunkSize);
-  }
-  return result;
-};
-
-function validateIpAndPort(input) {
-  var parts = input.split(":");
-  var ip = parts[0].split(".");
-  var port = parts[1];
-  return validateNum(port, 1, 65535) && ip.length == 4 && ip.every(function (segment) {
-    return validateNum(segment, 0, 255);
-  });
-}
-
-function validateNum(input, min, max) {
-  var num = +input;
-  return num >= min && num <= max && input === num.toString();
-}
 
 //  -------------------------
 
@@ -93,43 +84,50 @@ bot.on('ready', async () => {
 
 //  -------------------------
 
-bot.on("messageCreate", async (message) => {
-  if (message.author.bot) return;
+for (const file of commandFiles) {
+  const command = require(`./commands/${file}`);
+  bot.commands.set(command.name, command);
+}
 
+bot.on('messageCreate', async (message) => {
+  if (!message.content.startsWith(PREFIX) || message.author.bot) return
+  const args = message.content.slice(PREFIX.length).trim().split(/ +/);
+  const command = args.shift().toLowerCase();
+
+  if (!bot.commands.has(command)) return
+  try {
+    bot.commands.get(command).execute(message, args);
+  } catch (error) {
+    console.error(error);
+  }
+
+});
+
+//  -------------------------
+
+bot.on('messageCreate', async (message) => {
   if (message.author.bot || !message.guild) return;
   let args = message.content.toLowerCase().split(" ");
   let command = args.shift()
 
-  if (command == PREFIX + `help`) {
-    let embed = new Discord.MessageEmbed()
-      .setTitle(`Bot commands list`)
-      .setDescription(`\`\`\`fix
-> ${PREFIX}s <name players> - Search players by name.
-> ${PREFIX}id <number id>   - Search players by number.
-> ${PREFIX}all              - Show all players.
-> ${PREFIX}ip <ip:port>     - Shows status of a given server.
-> ${PREFIX}start            - Send status server to channel.
-> ${PREFIX}stop             - Stop send status server to channel.
-> ${PREFIX}stopbot          - Used if your bot server doesn't have a manual stop system.
-> ${PREFIX}clear <number>   - Clear all message from bots\`\`\``)
-      .setTimestamp()
-      .setColor(COLORBOX)
-      .setFooter({ text: `Github: Kuju29/fivem-bots-discord` })
-    message.reply({ embeds: [embed] }).then((msg) => {
-      console.log(`Completed ${PREFIX}help`);
-      setTimeout(() => {
-        if (AUTODELETE) {
-          msg.delete();
-          console.log(`Auto delete message ${PREFIX}help`);
-        }
-      }, 10000);
-    });
-  }
-
-  // ----------------------------------------------
-
   if (command == PREFIX + 'start') {
     console.log(`Completed ${PREFIX}start`);
+    if (config.NCOMMAND) {
+      let embedss = new Discord.MessageEmbed()
+        .setColor(COLORBOX)
+        .setDescription(`Completed \`${PREFIX}start\``)
+      message.reply({
+        embeds: [embedss]
+      }).then((msg) => {
+
+        setTimeout(() => {
+          if (AUTODELETE) {
+            msg.delete();
+            console.log(`Delete notification message ${PREFIX}start`);
+          }
+        }, 5000);
+      });
+    }
     sTart = setInterval(async function () {
       inFo.checkOnlineStatus().then(async (server) => {
         if (server) {
@@ -196,215 +194,22 @@ bot.on("messageCreate", async (message) => {
   if (command == PREFIX + 'stop') {
     clearInterval(sTart);
     console.log(`Completed ${PREFIX}stop`);
-  }
-
-  // ---------------------------------------------- 
-  
-  inFo.getPlayers().then(async (players) => {
-    if (command == PREFIX + 's') {
-      let text = message.content.toLowerCase().substr(3, 20);
-      let playerdata = players.filter(function (person) { return person.name.toLowerCase().includes(`${text}`) });
-      let result1 = [];
-      let index = 1;
-      for (let player of playerdata) {
-        result1.push(`${index++}. ${player.name} | ID : ${player.id} | Ping : ${player.ping}\n`);
-      };
-      const result = result1.join("\n").toString();
-      let embed = new Discord.MessageEmbed().setTimestamp();
-      if (message.member.permissions.has(PERMISSION)) {
-        embed.setColor(COLORBOX)
-          .setTitle(`Search player | ${SERVER_NAME}`)
-          .setDescription(result.length > 0 ? result : 'No Players')
-        message.reply({ embeds: [embed] }).then((msg) => {
-          console.log(`Completed ${PREFIX}s ${text}`);
-          setTimeout(() => {
-            if (AUTODELETE) {
-              msg.delete();
-              console.log(`Auto delete message ${PREFIX}s ${text}`);
-            }
-          }, 10000);
-        });
-      } else {
-        embed.setColor(COLORBOX)
-          .setTitle(`Search player | Error`)
-          .setDescription(`❌ You do not have the ${PERMISSION}, therefor you cannot run this command!`)
-        message.reply({ embeds: [embed] }).then((msg) => {
-          console.log(`Error ${PREFIX}s message`);
-          setTimeout(() => {
-            if (AUTODELETE) {
-              msg.delete();
-              console.log(`Auto delete Error message ${PREFIX}s message`);
-            }
-          }, 10000);
-        });
-      }
-    }
-
-    if (command == PREFIX + 'id') {
-      let num = message.content.match(/[0-9]/g).join('').valueOf();
-      let playerdata = players.filter(players => players.id == num);
-      let result1 = [];
-      let index = 1;
-      for (let player of playerdata) {
-        result1.push(`${index++}. ${player.name} | ID : ${player.id} | Ping : ${player.ping}\n`);
-      };
-      const result = result1.join("\n").toString();
-      let embed = new Discord.MessageEmbed().setTimestamp();
-      if (message.member.permissions.has(PERMISSION)) {
-        embed.setColor(COLORBOX)
-          .setTitle(`Search player | ${SERVER_NAME}`)
-          .setDescription(result.length > 0 ? result : 'No Players')
-        message.reply({ embeds: [embed] }).then((msg) => {
-          console.log(`Completed ${PREFIX}id ${num}`);
-          setTimeout(() => {
-            if (AUTODELETE) {
-              msg.delete();
-              console.log(`Auto delete message ${PREFIX}id ${num}`);
-            }
-          }, 10000);
-        });
-      } else {
-        embed.setColor(COLORBOX)
-          .setTitle(`Search player | Error`)
-          .setDescription(`❌ You do not have the ${PERMISSION}, therefor you cannot run this command!`)
-        message.reply({ embeds: [embed] }).then((msg) => {
-          console.log(`Error ${PREFIX}id message`);
-          setTimeout(() => {
-            if (AUTODELETE) {
-              msg.delete();
-              console.log(`Auto delete message Error ${PREFIX}id message`);
-            }
-          }, 10000);
-        });
-      }
-    }
-
-    if (command == PREFIX + 'all') {
-      let result = [];
-      let index = 1;
-      for (let player of players) {
-        result.push(`${index++}. ${player.name} | ID : ${player.id} | Ping : ${player.ping}\n`);
-      };
-      if (message.member.permissions.has(PERMISSION)) {
-        let chunks = splitChunks(result.join("\n").toString(), 2000);
-        // let chunks = Discord.Util.splitMessage(result.join("\n"))
-        let embed = new Discord.MessageEmbed().setTitle(`All_players | ${SERVER_NAME}`);
-        if (result.length > 1) {
-          const embeds = chunks.map((chunk) => {
-            return new Discord.MessageEmbed()
-              .setColor(COLORBOX)
-              .setDescription(chunk)
-          });
-          await new Pagination(message.channel, embeds, "Part").paginate();
-          console.log(`Completed !all`);
-        } else {
-          embed.setColor(COLORBOX)
-            .setDescription(result.length > 0 ? result : 'No Players')
-          message.reply({ embeds: [embed] }).then((msg) => {
-            console.log(`Completed ${PREFIX}all No Players`);
-            setTimeout(() => {
-              if (AUTODELETE) {
-                msg.delete();
-                console.log(`Auto delete message ${PREFIX}all No Players`);
-              }
-            }, 10000);
-          });
-        }
-      } else {
-        let embed = new Discord.MessageEmbed()
-          .setColor(COLORBOX)
-          .setTitle(`Search player | Error`)
-          .setDescription(`❌ You do not have the ${PERMISSION}, therefor you cannot run this command!`)
-          .setTimestamp(new Date());
-        message.reply({ embeds: [embed] }).then((msg) => {
-          console.log(`Error ${PREFIX}all`);
-          setTimeout(() => {
-            if (AUTODELETE) {
-              msg.delete();
-              console.log(`Auto delete message Error ${PREFIX}all`);
-            }
-          }, 10000);
-        });
-      }
-    }
-
-  }).catch((err) => {
-    console.log(`Catch ERROR or Offline: ` + err);
-  });
-
-  if (command == PREFIX + 'ip') {
-    let text = message.content.toLowerCase().substr(4, 24);
-    let testip = validateIpAndPort(text);
-    const iNfo = new fivem.ApiFiveM(text);
-    if (testip) {
-      iNfo.checkOnlineStatus().then(async (server) => {
-        if (server) {
-          let infoplayers = (await iNfo.getDynamic());
-          let embed = new Discord.MessageEmbed()
-            .setColor(COLORBOX)
-            .setTitle(`Server: \`${text}\``)
-            .addField('**Server Status**', `\`\`\`✅Online\`\`\``, true)
-            .addField('**Online Players**', `\`\`\`${infoplayers.clients}/${infoplayers.sv_maxclients}\`\`\``, true)
-            .setTimestamp(new Date());
-          message.reply({ embeds: [embed] }).then((msg) => {
-            console.log(`Completed ${PREFIX}ip ${text} online`);
-            setTimeout(() => {
-              if (AUTODELETE) {
-                msg.delete();
-                console.log(`Auto delete message ${PREFIX}ip ${text} online`);
-              }
-            }, 10000);
-          });
-        } else {
-          let embed = new Discord.MessageEmbed()
-            .setColor(COLORBOX)
-            .setTitle(`Server: \`${text}\``)
-            .addField('**Server Status**', `\`\`\`❌Offline or Invalid IP\`\`\``, true)
-            .addField('**Online Players**', `\`\`\`-/-\`\`\``, true)
-            .setTimestamp(new Date());
-          message.reply({ embeds: [embed] }).then((msg) => {
-            console.log(`Completed ${PREFIX}ip ${text} offline`);
-            setTimeout(() => {
-              if (AUTODELETE) {
-                msg.delete();
-                console.log(`Auto delete message ${PREFIX}ip ${text} offline`);
-              }
-            }, 10000);
-          });
-        }
-      }).catch((err) => {
-        console.log(`Catch ERROR or Offline: ` + err);
-      });
-    } else {
-      let embed = new Discord.MessageEmbed()
+    if (config.NCOMMAND) {
+      let embedss = new Discord.MessageEmbed()
         .setColor(COLORBOX)
-        .addField(`**Are you sure the IP is correct?**`, `\`${text}\``, true)
-        .setTimestamp(new Date());
-      message.reply({ embeds: [embed] }).then((msg) => {
-        console.log(`Completed ${PREFIX}ip Check IP: ${text}`);
-        setTimeout(() => {
-          if (AUTODELETE) {
-            msg.delete();
-            console.log(`Auto delete message ${PREFIX}ip Are you sure the IP is correct? ${text}`);
-          }
-        }, 10000);
-      });
-    };
-  }
+        .setDescription(`Completed \`${PREFIX}stop\``)
+      if (config.NCOMMAND) return message.reply({
+        embeds: [embedss]
+      }).then((msg) => {
 
-  if (command == PREFIX + 'clear' && AUTODELETE == false) {
-    let num = message.content.match(/[0-9]/g).join('').valueOf();
-    const Channel = message.channel;
-    const Messages = await Channel.messages.fetch({ limit: num });
-    Messages.forEach(message => {
-      if (message.author.bot) message.delete()
-    });
-    console.log(`Completed ${PREFIX}Clear ${num}`);
-  }
-  
-  if (command == PREFIX + 'stopbot') {
-    console.log(`${PREFIX}stopbot - the bot has been stopped.....`)
-    process.exit(0);
+        setTimeout(() => {
+
+          msg.delete();
+          console.log(`Delete notification message ${PREFIX}stop`);
+
+        }, 5000);
+      });
+    }
   }
 
 });
